last fixe

+++ Blog/blog_delete.php
@@ -15,9 +15,9 @@
         if(!empty($_GET['eID'])){
             $eID = $_GET['eID'];
         }
-        if(!empty($_GET['duID'])){
-            $duID = $_GET['duID'];
-            $fID = $_GET['duID'];
+        if(!empty($_GET['euID'])){
+            $euID = $_GET['euID'];
+            $fID = $_GET['euID'];
         }
 
 
@@ -31,6 +31,17 @@
                 $lock= $row["locked"];
             }
         } 
+
+        $stmt = $conn->prepare("SELECT blog_entry.uID FROM blog_entry WHERE ID = ?");  //gets if the user is locked
+        $stmt->bind_param("i", $eID);   
+        $stmt->execute(); 
+        $result = $stmt->get_result(); 
+
+        if ($result->num_rows > 0) {
+            while($row = $result->fetch_assoc()) {    
+                $cID= $row["uID"];
+            }
+        } 
         
 
 
@@ -42,18 +53,19 @@
 
         if ($res["userType"] == "endUser"){
             if ($lock == 0){
-                if ($eID != 0){    
-                    $stmt = $conn->prepare("DELETE FROM blog_entry WHERE blog_entry.ID = ? AND blog_entry.uID = ?");  // deletes entries by specific id
-                    $stmt->bind_param("ii", $eID, $uID); 
-                    $stmt->execute();  
-                    $result = $stmt->get_result(); 
-                    echo $result;
-                    if ($result == 0){
+                if ($eID != 0){
+                    if ($cID == $uID){
+                        $stmt = $conn->prepare("DELETE FROM blog_entry WHERE blog_entry.ID = ? AND blog_entry.uID = ?");  // deletes entries by specific id
+                        $stmt->bind_param("ii", $eID, $uID); 
+                        $stmt->execute();  
+                
                         $data = ["Action"=>"Entry deleted"];
+                        jsonWrite($version,$data);  
+                    }    
+                    else{
+                        $data = ["Action"=>"you don't have permission you big dum"];
                         jsonWrite($version,$data);
-                    }
-                    $data = ["Action"=>"you don't have permission you big dum"];
-                    jsonWrite($version,$data);
+                    }   
                 }    
                 else{
                     errorWrite($version,"Wrong inputs");
@@ -66,13 +78,13 @@
         }
         else if ($res["userType"] == "admin"){
             if ($lock == 0){
-                if ($duID != 0){    
+                if ($euID != 0){    
                     $stmt = $conn->prepare("DELETE FROM blog WHERE blog.uID = ?"); // deletes blogs by specific id
-                    $stmt->bind_param("i", $duID); 
+                    $stmt->bind_param("i", $euID); 
                     $stmt->execute(); 
 
                     $stmt = $conn->prepare("DELETE FROM blog_entry WHERE blog_entry.uID = ?"); // deletes entries from the blog when deleted
-                    $stmt->bind_param("i", $duID); 
+                    $stmt->bind_param("i", $euID); 
                     $stmt->execute(); 
                     
                     $data = ["Action"=>"Blog deleted"];
@@ -97,6 +109,6 @@
 
 
     // ?eID=x
-    // ?duID=x
+    // ?euID=x
 ?>
